pyproject.toml
[tool.poetry]
name = "fin-assistant"
version = "0.2.0"
description = "Agentic LLM assistant with clean structure"
authors = ["Vlad Kolinko <vladkolinko@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.0"
uvicorn = "^0.29.0"
openai = "^1.23.2"
python-dotenv = "^1.0.1"
tiktoken = "^0.9.0"
redis = "^6.2.0"
pydantic-settings = "^2.9.1"
python-json-logger = "^3.3.0"
tenacity = "^9.1.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
